requirements.txt
fastapi
uvicorn
scikit-learn
xgboost
imbalanced-learn
python-dotenv
matplotlib
seaborn
python-multipart
pydantic